+++ devicemanagement/tests/TestDevMgmtApiSyncCmds.cpp
@@ -345,41 +345,33 @@ static inline bool decodeSingleTraceEvent(std::stringstream& logs, const struct 
   return validTypeFound;
 }
 
-static inline uint32_t updatePmicFwMetaVersion(std::string pmicSlotImgPath, uint32_t& oldVersion,
-                                               uint32_t& newVersion) {
+static inline uint32_t updatePmicMetaFwHash(std::string pmicSlotImgPath, char* newHash, char* oldHash) {
   std::fstream pmicImage;
   // Open the PMIC slot image
   pmicImage.open(pmicSlotImgPath, std::fstream::binary | std::fstream::in | std::fstream::out);
   if (pmicImage.is_open() && (pmicImage.tellp() <= 0)) {
     int pmicMetaStartOffset = kFlashImageRawFileHeaderSize + kPmicFlashImageMetadataOffset;
-    // Seek to the start of the PMIC FW version in metadata
-    pmicImage.seekg(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, version), std::fstream::beg);
-    // Read the PMIC FW version
-    pmicImage.read(templ::bit_cast<char*>(&oldVersion), sizeof(oldVersion));
-    DV_LOG(INFO) << "PMIC current version: " << EXTRACT_BYTE(2, oldVersion) << "." << EXTRACT_BYTE(1, oldVersion) << "."
-                 << EXTRACT_BYTE(0, oldVersion);
-
-    // Increment the patch version for testing
-    uint8_t newPatch = EXTRACT_BYTE(0, oldVersion) + 1;
-    newVersion = PMIC_FORMAT_VERSION(EXTRACT_BYTE(2, oldVersion), EXTRACT_BYTE(1, oldVersion), newPatch);
-    DV_LOG(INFO) << "PMIC new version: " << EXTRACT_BYTE(2, newVersion) << "." << EXTRACT_BYTE(1, newVersion) << "."
-                 << EXTRACT_BYTE(0, newVersion);
-
-    // Seek to the start of the PMIC FW version in metadata
-    pmicImage.seekp(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, version), std::fstream::beg);
-    // Write the new PMIC FW version
-    pmicImage.write(templ::bit_cast<char*>(&newVersion), sizeof(newVersion));
+    // Seek to the start of the PMIC FW hash in metadata
+    pmicImage.seekg(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, hash), std::fstream::beg);
+    // Read the PMIC FW hash
+    pmicImage.read(oldHash, 16);
+    DV_LOG(INFO) << "PMIC current hash: " << oldHash;
+    DV_LOG(INFO) << "PMIC new hash: " << newHash;
+    // Seek to the start of the PMIC FW hash in metadata
+    pmicImage.seekp(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, hash), std::fstream::beg);
+    // Write the new PMIC FW hash
+    pmicImage.write(newHash, 16);
 
     // Re-compute the checksum
     uint32_t checksum = 0;
     // Write zero initially to checksum
     pmicImage.seekp(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, checksum), std::fstream::beg);
-    // Write the new PMIC FW version
+    // Write the new PMIC FW checksum
     pmicImage.write(templ::bit_cast<char*>(&checksum), sizeof(checksum));
     // Get the total size of PMIC image
     uint32_t fileSize;
     pmicImage.seekg(pmicMetaStartOffset + offsetof(pmicRuntimeMetadata_t, image_size), std::fstream::beg);
-    // Read the PMIC FW version
+    // Read the PMIC FW size
     pmicImage.read(templ::bit_cast<char*>(&fileSize), sizeof(fileSize));
 
     // Now recompute the checksum from the beginning of PMIC image
@@ -2914,22 +2906,18 @@ void TestDevMgmtApiSyncCmds::setFirmwareUpdateImage(bool singleDevice, bool rese
   std::string pmicSlot1ImgPath = DEVICE_SIGNED_ARTIFACTS_PATH + std::string("/signed_pmic_s1_fw.img");
   std::string flashToolCmd =
     FLASH_TOOL_PATH + std::string(" create ") + FLASH_IMG_PATH + std::string(" ") + FLASH_32MBIT_TEMPLATE_PATH;
+  char testHash1[16] = "dead";
+  char testHash2[16] = "beef";
 
   auto deviceCount = singleDevice ? 1 : dm.getDevicesCount();
-
   for (int iter = 0; iter < iterations; iter++) {
-    uint32_t s0OldVersion = 0;
-    uint32_t s0NewVersion = 0;
-    uint32_t s1OldVersion = 0;
-    uint32_t s1NewVersion = 0;
-    // Increment the pmic slot 0 fw version
-    ASSERT_TRUE(updatePmicFwMetaVersion(pmicSlot0ImgPath, s0OldVersion, s0NewVersion));
-    // Increment the pmic slot 1 fw version
-    ASSERT_TRUE(updatePmicFwMetaVersion(pmicSlot1ImgPath, s1OldVersion, s1NewVersion));
-
-    // The slot 0 and slot 1 version should match
-    ASSERT_EQ(s0OldVersion, s1OldVersion);
-    ASSERT_EQ(s0NewVersion, s1NewVersion);
+    char s0OldHash[16];
+    char s1OldHash[16];
+    char* testHashPtr = (iter % 2) ? testHash1 : testHash2;
+    // Change the pmic slot 0 fw hash
+    ASSERT_TRUE(updatePmicMetaFwHash(pmicSlot0ImgPath, testHashPtr, s0OldHash));
+    // Change the pmic slot 1 fw hash
+    ASSERT_TRUE(updatePmicMetaFwHash(pmicSlot1ImgPath, testHashPtr, s1OldHash));
 
     // Run command to invoke flash tool to re-create the flash image
     ASSERT_EQ(system(flashToolCmd.c_str()), 0);
@@ -2953,36 +2941,14 @@ void TestDevMgmtApiSyncCmds::setFirmwareUpdateImage(bool singleDevice, bool rese
         ASSERT_EQ(output_buff[0], 0);
       }
 
-      const uint32_t output_size_v = sizeof(device_mgmt_api::firmware_version_t);
-      char output_buff_v[output_size_v] = {0};
-
       // Optionally reset the device
       if (resetDev) {
         ASSERT_EQ(dm.serviceRequest(deviceIdx, device_mgmt_api::DM_CMD::DM_CMD_RESET_ETSOC, nullptr, 0, nullptr, 0,
                                     hst_latency.get(), dev_latency.get(), DURATION2MS(end - Clock::now())),
                   device_mgmt_api::DM_STATUS_SUCCESS);
         DV_LOG(INFO) << "Service Request Completed for Device: " << deviceIdx;
-
-        // Check the PMIC FW version to match to new one. Its okay to match either of the
-        // slot version since both are same
-        ASSERT_EQ(dm.serviceRequest(deviceIdx, device_mgmt_api::DM_CMD::DM_CMD_GET_MODULE_FIRMWARE_REVISIONS, nullptr,
-                                    0, output_buff_v, output_size_v, hst_latency.get(), dev_latency.get(),
-                                    DURATION2MS(end - Clock::now())),
-                  device_mgmt_api::DM_STATUS_SUCCESS);
-        DV_LOG(INFO) << "Service Request Completed for Device: " << deviceIdx;
       }
-      // Validate FW version
-      if (resetDev && (getTestTarget() != Target::Loopback)) {
-        device_mgmt_api::firmware_version_t* firmware_versions = (device_mgmt_api::firmware_version_t*)output_buff_v;
-        DV_LOG(INFO) << fmt::format("\tCurrent PMIC version: {}.{}.{}", MAJOR_VERSION(firmware_versions->pmic_v),
-                                    MINOR_VERSION(firmware_versions->pmic_v),
-                                    REVISION_VERSION(firmware_versions->pmic_v));
-
-        // Match the PMIC version
-        ASSERT_EQ(MAJOR_VERSION(firmware_versions->pmic_v), EXTRACT_BYTE(2, s0NewVersion));
-        ASSERT_EQ(MINOR_VERSION(firmware_versions->pmic_v), EXTRACT_BYTE(1, s0NewVersion));
-        ASSERT_EQ(REVISION_VERSION(firmware_versions->pmic_v), EXTRACT_BYTE(0, s0NewVersion));
-      }
+      // TODO: Validate the hash once the ability to read PMIC has is available
     }
   }
 }